pyproject.toml
[project]
name = "finetune-embeddings"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.6.1",
    "openai>=2.26.0",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.2",
]
